Added methods for user registration , automatic community creation and wallet generation

+++ src/main/java/com/educare/unitylend/model/Wallet.java
@@ -12,5 +12,5 @@ import java.util.UUID;
 public class Wallet {
     private String walletId;
     private User user;
-    private BigDecimal balance;
+    private Float balance;
 }
